+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/Sample.java
@@ -48,7 +48,7 @@ import uk.ac.bbsrc.tgac.miso.core.security.SecurableByProfile;
  */
 public interface Sample
     extends SecurableByProfile, Locatable, Comparable<Sample>, Boxable, ChangeLoggable, Aliasable, Serializable,
-    QualityControllable<SampleQC>, Deletable {
+    QualityControllable<SampleQC>, Deletable, Attachable {
 
   /** Field UNSAVED_ID */
   public static final long UNSAVED_ID = 0L;