+++ miso-web/src/main/webapp/scripts/hot_sample.js
@@ -545,7 +545,11 @@ HotTarget.sample = (function() {
             description: 'The date that the sample was created.',
             include: Constants.isDetailedSample && !config.isLibraryReceipt,
             unpack: function(sam, flat, setCellMeta) {
-              flat.creationDate = Utils.valOrNull(sam.creationDate);
+              if (!sam.creationDate && config.propagate) {
+                flat.creationDate = Utils.getCurrentDate();
+              } else {
+                flat.creationDate = Utils.valOrNull(sam.creationDate);
+              }
             },
             pack: function(sam, flat, errorHandler) {
               sam.creationDate = flat.creationDate;
@@ -988,8 +992,11 @@ HotTarget.sample = (function() {
           HotUtils.makeChildren('sample', HotUtils.relationCategoriesForDetailed().concat(
               [HotUtils.relations.library(), HotUtils.relations.dilution(), HotUtils.relations.pool()]))].concat(
           HotUtils.makeQcActions("Sample")).concat(
-          config.worksetId ? [HotUtils.makeRemoveFromWorkset('samples', config.worksetId)] : [HotUtils.makeAddToWorkset('samples',
-              'sampleIds')]);
+          [
+              config.worksetId ? HotUtils.makeRemoveFromWorkset('samples', config.worksetId) : HotUtils.makeAddToWorkset('samples',
+                  'sampleIds'), HotUtils.makeAttachFile('sample', function(sample) {
+                return sample.projectId;
+              })]);
     },
 
     getCustomActions: function(table) {