+++ miso-web/src/main/webapp/scripts/hot_tissuematerial.js
@@ -30,6 +30,7 @@ HotTarget.tissuematerial = {
             var deleteNext = function(index) {
               if (index == items.length) {
                 window.location = window.location.origin + '/miso/tissuematerial/list';
+                return;
               }
               Utils.ajaxWithDialog('Deleting ' + items[index].alias, 'DELETE', '/miso/rest/tissuematerial/' + items[index].id, null,
                   function() {