+++ miso-web/src/main/webapp/scripts/list_attachment.js
@@ -22,73 +22,6 @@
  */
 
 ListTarget.attachment = (function() {
-  var showUploadDialog = function(entityType, entityId) {
-    var dialogArea = jQuery('#dialog');
-    dialogArea.empty();
-
-    var form = jQuery('<form id="uploadForm">');
-    form.append(jQuery('<input id="fileInput" type="file" name="file">'));
-    dialogArea.append(form);
-
-    var dialog = dialogArea.dialog({
-      autoOpen: true,
-      width: 500,
-      title: 'Attach File',
-      modal: true,
-      buttons: {
-        upload: {
-          id: 'upload',
-          text: 'Upload',
-          click: function() {
-            if (!jQuery('#fileInput').val()) {
-              alert('No file selected!');
-              return;
-            }
-            var formData = new FormData(jQuery('#uploadForm')[0]);
-            dialog.dialog("close");
-            dialogArea.empty();
-            dialogArea.append(jQuery('<p>Uploading...</p>'));
-
-            dialog = jQuery('#dialog').dialog({
-              autoOpen: true,
-              height: 400,
-              width: 350,
-              title: 'Uploading File',
-              modal: true,
-              buttons: {},
-              closeOnEscape: false,
-              open: function(event, ui) {
-                jQuery(this).parent().children().children('.ui-dialog-titlebar-close').hide();
-              }
-            });
-
-            jQuery.ajax({
-              url: '/miso/attachments/' + entityType + '/' + entityId,
-              type: 'POST',
-              data: formData,
-              cache: false,
-              contentType: false,
-              processData: false
-            }).success(function(data) {
-              dialog.dialog("close");
-              Utils.page.pageReload();
-            }).fail(function(xhr, textStatus, errorThrown) {
-              dialog.dialog("close");
-              Utils.showAjaxErrorDialog(xhr, textStatus, errorThrown);
-            });
-          }
-        },
-        cancel: {
-          id: 'cancel',
-          text: 'Cancel',
-          click: function() {
-            dialog.dialog("close");
-          }
-        }
-      }
-    });
-  };
-
   return {
     name: "Attachments",
     createUrl: function(config, projectId) {
@@ -100,8 +33,20 @@ ListTarget.attachment = (function() {
     createStaticActions: function(config, projectId) {
       return [{
         name: 'Upload',
-        handler: function() {
-          showUploadDialog(config.entityType, config.entityId);
+        handler: !config.projectId ? function() {
+          ListTarget.attachment.showUploadDialog(config.entityType, config.entityId);
+        } : function() {
+          Utils.showWizardDialog('Attach File', [{
+            name: 'Upload new file',
+            handler: function() {
+              ListTarget.attachment.showUploadDialog(config.entityType, config.entityId);
+            }
+          }, {
+            name: 'Link project file',
+            handler: function() {
+              ListTarget.attachment.showLinkDialog(config.entityType, config.entityId, config.projectId);
+            }
+          }]);
         }
       }];
     },
@@ -117,6 +62,15 @@ ListTarget.attachment = (function() {
                 return '<a href="/miso/attachments/' + config.entityType + '/' + config.entityId + '/' + full.id + '">' + data + '</a>';
               }
               return data;
+            }
+          },
+          {
+            sTitle: 'Category',
+            mData: 'category',
+            include: true,
+            iSortPriority: 0,
+            mRender: function(data, type, full) {
+              return data ? data : 'Misc.';
             }
           },
           {
@@ -153,6 +107,121 @@ ListTarget.attachment = (function() {
               Utils.page.pageReload();
             });
           });
+    },
+    showUploadDialog: function(entityType, entityId, sharedIds) {
+      var dialogArea = jQuery('#dialog');
+      dialogArea.empty();
+
+      var form = jQuery('<form id="uploadForm">');
+      form.append(jQuery('<p><input id="fileInput" type="file" name="file"></p>'));
+      var select = jQuery('<select id="attachmentCategory" type="text" name="attachmentCategory">');
+      [{
+        id: 0,
+        alias: 'Misc.'
+      }].concat(Constants.attachmentCategories).forEach(function(category) {
+        select.append(jQuery('<option value="' + category.id + '">' + category.alias + '</option>'));
+      });
+      var p = jQuery('<p>').append(jQuery('<label for="attachmentCategory">Category:</label>')).append(select);
+      form.append(p);
+      dialogArea.append(form);
+
+      var dialog = dialogArea.dialog({
+        autoOpen: true,
+        width: 500,
+        title: 'Attach File',
+        modal: true,
+        buttons: {
+          upload: {
+            id: 'upload',
+            text: 'Upload',
+            click: function() {
+              if (!jQuery('#fileInput').val()) {
+                alert('No file selected!');
+                return;
+              }
+              var params = {};
+              if (jQuery('#attachmentCategory').val() > 0) {
+                params.categoryId = jQuery('#attachmentCategory').val();
+              }
+              if (sharedIds) {
+                params.entityIds = sharedIds;
+              }
+              var url = '/miso/attachments/' + entityType + '/' + entityId + '?' + jQuery.param(params);
+              var formData = new FormData(jQuery('#uploadForm')[0]);
+              dialog.dialog("close");
+              dialogArea.empty();
+              dialogArea.append(jQuery('<p>Uploading...</p>'));
+
+              dialog = jQuery('#dialog').dialog({
+                autoOpen: true,
+                height: 400,
+                width: 350,
+                title: 'Uploading File',
+                modal: true,
+                buttons: {},
+                closeOnEscape: false,
+                open: function(event, ui) {
+                  jQuery(this).parent().children().children('.ui-dialog-titlebar-close').hide();
+                }
+              });
+
+              jQuery.ajax({
+                url: url,
+                type: 'POST',
+                data: formData,
+                cache: false,
+                contentType: false,
+                processData: false
+              }).success(function(data) {
+                dialog.dialog("close");
+                Utils.page.pageReload();
+              }).fail(function(xhr, textStatus, errorThrown) {
+                dialog.dialog("close");
+                Utils.showAjaxErrorDialog(xhr, textStatus, errorThrown);
+              });
+            }
+          },
+          cancel: {
+            id: 'cancel',
+            text: 'Cancel',
+            click: function() {
+              dialog.dialog("close");
+            }
+          }
+        }
+      });
+    },
+    showLinkDialog: function(entityType, entityId, projectId, sharedIds) {
+      var url = '/miso/rest/project/' + projectId + '/files';
+      Utils.ajaxWithDialog('Retrieving Project Files', 'GET', url, null, function(attachments) {
+        if (!attachments.length) {
+          Utils.showOkDialog('Link Project File', ['Project has no attachments to link.']);
+          return;
+        }
+
+        var attachmentIds = attachments.map(Utils.array.getId);
+        Utils.showDialog('Link Project File', 'Link', [{
+          label: 'File',
+          type: 'select',
+          required: true,
+          values: attachmentIds,
+          getLabel: function(value) {
+            return Utils.array.findUniqueOrThrow(Utils.array.idPredicate(value), attachments).filename;
+          },
+          property: 'attachmentId'
+        }], function(output) {
+          var params = {
+            fromEntityType: 'project',
+            fromEntityId: projectId,
+            attachmentId: output.attachmentId
+          }
+          if (sharedIds) {
+            params.entityIds = sharedIds;
+          }
+          var url = '/miso/rest/attachments/' + entityType + '/' + entityId + '?' + jQuery.param(params);
+          Utils.ajaxWithDialog('Linking File', 'POST', url, null, Utils.page.pageReload);
+        });
+      });
     }
   };
 })();